### 2025-06-24 #8462 @editorialbot
[PRE REVIEW]: HydroGenerate: A Python Package for Estimating the Technical Hydropower Potential
**Submitting author:** @snag427 (<a href="http://orcid.org/0000-0001-7571-4137">Soumyadeep Nag</a>)
**Repository:** https://github.com/IdahoLabResearch/HydroGenerate
**Branch with paper.md** (empty if default branch): Soumyadeep_Paper
**Version:** v1.0.0
