### 2025-07-24 #8462 @kyleniemeyer
@snag427 your paper remains in our queue - unfortunately it's a matter of editors becoming available after current assignments finish. We may have someone available now, though.
