### 2025-07-24 #8462 @AdamRJensen
There's a citation with 'Pacific Northwest National Laboratory' as the author where it is formatted as a name. Please enclose this in curly brackets. Also, there's a few typos that could be fixed by running a spell checker before the start of the review.
