### 2025-08-03 #8462 @AdamRJensen
> Hi [@AdamRJensen](https://github.com/AdamRJensen), I'm on holiday right now, but I am available to review from 25th August if that is okay? Looks like an interesting package that I would be happy to review.

No problem at all. Thanks for being willing to review!
