### 2025-08-19 d309f37
Paper update
Corrected typos
