+++ paper.md
@@ -27,35 +27,35 @@ bibliography: paper.bib
 
 # Summary
 
-Conducting pre-feasibility studies on hydro power sites can be expensive, technically challenging, and time-consuming for a variety of stakeholders from developers to small asset owners. HydroGenerate is an open-source Python library designed to assist hydropower developers and other stakeholders in estimating the generation potential at a user-selected sites, while also aiming to reduce the technical and financial burdens associated with the initial pre-feasibility analysis phase. HydroGenerate can help reduce the above mentioned barriers by providing realistic insights into a site's technical potential and performance. Thereby, helping hydropower stakeholders make decisions confidently and move projects forward if feasible. Additionally, HydroGenerate can support researchers in evaluating the potential of new hydropower technologies or plant operating policies. Also, due to the light-weight nature of the tool, HydroGenerate can be used alongside third-party packages to investigate solutions to  a variety of other energy system challenges (e.g., capacity expansion models, hybrid plant optimization, etc.). The tool is able to process historical site data (where available), estimate the technical potential, and provide a preliminary economic analysis. Both the tool and the accompanying user guide are crafted to facilitate ease of use for beginners, while also offering deeper analysis capabilities for more experienced users. 
+Conducting pre-feasibility studies on hydro power sites can be expensive, technically challenging, and time-consuming for a variety of stakeholders from developers to small asset owners. HydroGenerate is an open-source Python library designed to assist hydropower developers and other stakeholders in estimating the generation potential at a user-selected site, while also aiming to reduce the technical and financial burdens associated with the initial pre-feasibility analysis phase. HydroGenerate can help reduce the above-mentioned barriers by providing realistic insights into a site's technical potential and performance. Thereby helping hydropower stakeholders make decisions confidently and move projects forward if feasible. Additionally, HydroGenerate can support researchers in evaluating the potential of new hydropower technologies or plant operating policies. Also, due to the light-weight nature of the tool, HydroGenerate can be used alongside third-party packages to investigate solutions to a variety of other energy system challenges (e.g., capacity expansion models, hybrid plant optimization, etc.). The tool can process historical site data (where available), estimate technical potential, and provide a preliminary economic analysis. Both the tool and the accompanying user guide are crafted to facilitate ease of use for beginners, while also offering deeper analysis capabilities for more experienced users. 
 
 # Statement of need
 
 HydroGenerate offers a modular, customizable, and well-documented [@HG] tool for hydropower potential estimation, pre-design, and evaluation of various project types. HydroGenerate is equipped with multiple features to support this task. For example, a) turbine technology selection based on hydraulic head and design flow, b) power output estimation using  efficiency curves that are specific to turbine type and size, c) compatibility with United States Geological Survey (USGS) dataretrieval Python package [@USGS] to minimize data management and user learning burdens, d) flow data processing for technology-based or maintenance constraints, e) head-loss calculation methods, f) embedded cost estimates based on the hydropower project type (i.e., non-powered dams, new stream-reach, canal-conduit, pumped storage hydro with existing infrastructure, new pumped storage hydro, unit additions to existing hydropower units, and generator rewinds) based on ORNL baseline cost model [@osti_1244193]. These features also distinguish HydroGenerate from closed and membership-based platforms such as RETScreen [@RETSCREEN], which lacks functionality for flow data analysis and operational constraints related to practical operation, head loss calculation in the penstock, and initial capital cost and operation and maintenance estimations. 
 
[…]
