### 2025-10-22 bba9f35
Addressing JOSS reviewer's comments.
This update addresses:
#25 by changing the Python version to 3.10, as suggested in #26. #25 will be fixed from a different branch.
#27 by adding two new subsections to the documentation. A new section in developer guide describes how to contribute and a new section in user guide describes how to seek support and report issues.
#28 by adding a note directing users to the location of the example data file in every instance where it is used.
#29 by removing a print statement that is not applicable for the example presented.
