+++ docs/UserGuide_8_Issues_Support.md
@@ -1,0 +1,41 @@
+# 🛠 Getting Help and Reporting Issues
+
+We want HydroGenerate to be reliable, useful, and easy to work with. This guide explains how to seek support and report issues effectively.
+
+---
+
+## How to Seek Support
+
+If you need help using HydroGenerate:
+
+- **Check the documentation**: Review the README and any available usage examples.
+- **Search existing issues**: Your question may already be answered in [GitHub Issues](https://github.com/IdahoLabResearch/HydroGenerate).
+- **Open a new discussion or issue** if you can't find what you need:
+  - Use the **"Question"** or **"Support"** label if available.
+  - Be clear and concise about what you're trying to do and where you're stuck.
+  - Include code snippets, error messages, and environment details when possible.
+
+For GitHub-related platform issues (e.g., account problems, repository access), contact [GitHub Support](https://docs.github.com/en/support/contacting-github-support) directly.
+
+---
+
+## How to Report Issues
+
+Bug reports and enhancement requests help us improve HydroGenerate. To report an issue:
+
+1. **Search existing issues** to avoid duplicates.
+2. If it’s new, [open an issue](https://github.com/IdahoLabResearch/HydroGenerate) and fill out the template completely:
+   - Describe the problem clearly.
+   - Include steps to reproduce it.
+   - Provide relevant logs, error messages, or screenshots.
+   - Mention your environment (OS, Python version, HydroGenerate version).
+
+Following these steps helps maintainers understand and resolve the issue faster.
+
+For more guidance, see:
+- [GitHub Docs: Creating an Issue](https://docs.github.com/en)
+- [pandas Contribution Guide](https://pandas.pydata.org/docs/development/contributing.html)
+
+---
+
+Thank you for helping make HydroGenerate better!
