+++ _sources/Theory_6_FlowConstraints.md
@@ -2,7 +2,7 @@
 To increase the accuracy of electricity generation estimates, certain real-world constraints such as, minimum and maximum flow constraints and annual and 5 yearly maintenance constraints, have been added to restrict the allowable flow through the turbine. Anuual maintenance and 5 yearly can be activated or deactivated using flags, <span style="color:blue"> 'annual_maintenance_flag' </span>  and <span style="color:blue">'major_maintenance_flag'</span>, respectively. 
 
 ### Flow Constraints for Design 
-The <span style="color:blue"> 'minimum_turbineflow' </span>  and <span style="color:blue"> 'minimum_turbineflow_percent' </span> parameters can be used to declare lower limit of flow, and can be either a cfs value (minimum_turbineflow) or a percentage of the design flow (minimum_turbineflow_percent). If neither of these parameters are defined by the user, Hydrogenerate assumes 10% of design flow as the minimum admissible flow for power generation. The maximum flow through the turbine is limited to the design flow (calculated internally or user-defined using <span style="color:blue">'design_flow'</span>). By default, the minimum and maximum flow constraints are always active.
+The <span style="color:blue"> 'minimum_turbineflow' </span> parameter, can be used to declare lower limit of flow, and can be either a cfs value or a percentage of the design flow. If the <span style="color:blue"> 'minimum_turbineflow' </span> parameter is not defined by the user, Hydrogenerate assumes 10% of design flow as the minimum admissible flow for power generation. The maximum flow through the turbine is limited to the design flow (calculated internally or user-defined using <span style="color:blue">'design_flow'</span>). By default, the minimum and maximum flow constraints are always active.
 
 ### Workflow
 Once design flow is calculated and turbine type is determined, the flow and maintenance constraints are executed in the following order:
