### 2025-10-28 23c43df
Docstring and test ideas
This commit adds doctring to the code and includes a new file with test ideas.
