+++ HydroGenerate/hydraulic_processing.py
@@ -1,12 +1,15 @@
-'''
-Copyright 2023, Battelle Energy Alliance, LLC
-'''
-
 """
 03-2023
 @author: Camilo J. Bastidas Pacheco
-This script handles head loss calculations and penstock sizing 
+This script handles head loss calculations and penstock sizing.
+
+Functionality
+-------------
+- Compute Reynolds number and select/compute friction factors.
+- Estimate head loss via Darcy–Weisbach or Hazen–Williams.
+- Size penstocks to respect a maximum allowable head loss.
 """
+
 
 import numpy as np
 from abc import ABC, abstractmethod
@@ -16,6 +19,50 @@ nu = 0.0000011223       # Kinematic viscosity of water (m2/s) at 60F (~15.6C) - 
 
 # Hydraulic design parameters
 class HydraulicDesignParameters:
+    """
+    Container for hydraulic inputs/derived values used in head-loss and sizing.
+
+    Parameters
+    ----------
+    flow : float or numpy.ndarray
+        Flow values (m^3/s) for general context or time-series use.
+    design_flow : float
+        Design discharge (m^3/s) used for diameter/head-loss sizing.
+    head : float or None
+        Available gross head (m).
+    head_loss : float or numpy.ndarray or None
+        Head loss (m). Can be precomputed or filled by calculators.
+    max_headloss_allowed : float or None
+        Maximum head loss as a percent of `head` (1–100). If None, 10% default
+        is applied in Darcy–Weisbach sizing.
+    penstock_headloss_method : str or None
+        Text identifier for method: e.g., "Darcy-Weisbach" or "Hazen-Williams".
+    penstock_headloss_calculation : bool
+        If True, compute penstock head loss; set False when net head is given.
+    penstock_length : float or None
+        Penstock length (m).
+    penstock_diameter : float or None
+        Known penstock diameter (m). If None, methods will size it.
+    penstock_material : str or None
+        Material key in roughness tables (e.g., "Steel", "Concrete", etc.).
+    penstock_frictionfactor : float or None
+        For Darcy–Weisbach: friction factor `f`. For Hazen–Williams: C-factor.
+        If None, it is computed/selected.
+    channel_average_velocity : float or None
+        Average channel velocity (m/s), if used elsewhere.
+
+    Attributes
+    ----------
+    Re : list or float
+        Reynolds number (dimensionless), computed for design flow.
+    penstock_design_diameter : float or None
+        Working/selected design diameter (m).
+    penstock_design_headloss : float or None
+        Head loss (m) at design conditions for current diameter.
+    channel_design_headloss : float or None
+    channel_headloss : float or None
+    """
+
 
     def __init__(self, flow, design_flow,       # Flow parameters
                  head, head_loss, 
@@ -50,12 +97,35 @@ class HydraulicDesignParameters:
 
     # Reynolds number calculation
     def reynoldsnumber_calculator(self):
+        """
+        Compute Reynolds number at design conditions.
+
+        Notes
+        -----
+        Uses :math:`Re = 4Q / (\\pi D \\nu)` with kinematic viscosity `nu`
+        defined at module level.
+
+        Side Effects
+        ------------
+        - Ensures `penstock_design_diameter` is initialized via
+        `designdiameter_calculator`.
+        - Updates `self.Re`.
+        """
 
         self.designdiameter_calculator()        # calculate diameter
         self.Re = 4 * self.design_flow / (np.pi * self.penstock_design_diameter * nu)        # Reynolds number, update  
 
     # Design diameter calculator - initial assumed diameter
     def designdiameter_calculator(self):
+        """
+        Initialize the working design diameter if needed.
+
+        Logic
+        -----
+        - If `penstock_diameter` is provided, use it as the design diameter.
[…]
