+++ HydroGenerate/summary_results.py
@@ -3,11 +3,39 @@ Copyright 2023, Battelle Energy Alliance, LLC
 '''
 
 """
-10-2024 
+10-2024
 @author: Camilo J. Bastidas Pacheco, J. Gallego-Calderon, Soumyadeep Nag
 
-This module includes function to summarize HydroGenerate results. 
+This module includes functions to summarize and visualize HydroGenerate results.
+
+Overview
+--------
+Provides post-processing utilities to:
+1) Plot turbine efficiency and power curves.
+2) Generate flow-duration, turbine-type, and monthly performance figures.
+3) Support both SI and US customary units.
+
+Expected Input Object
+---------------------
+All plotting functions receive an object `x` produced by HydroGenerate,
+which typically exposes:
+- pandas_dataframe : bool
+- dataframe_output : pandas.DataFrame
+- flowduration_curve : pandas.DataFrame (optional)
+- design_flow : float
+- pctime_runfull : float
+- rated_power : float
+- net_head : float
+- units : {'US', 'SI'}
+
+Notes
+-----
+- Figures are returned as `matplotlib.figure.Figure` objects and closed
+  (`plt.close()`) after creation.
+- Functions print a short message instead of raising an error when
+  required attributes are missing.
 """
+
 
 import numpy as np
 import pandas as pd
@@ -21,7 +49,30 @@ cfs_to_cms = 0.0283168 # 1 cubic feet per second to cubic meter per second
 ft_to_m = 0.3048 # 1 feet to meters
 
 # function to plot turbine efficiency and power generation as a function of flow
-def flow_efficiency_power_plot(x) :
+def flow_efficiency_power_plot(x):
+    """
+    Plot turbine efficiency and power (kW) as functions of discrete turbine flow.
+
+    Parameters
+    ----------
+    x : object
+        HydroGenerate results container with:
+        - pandas_dataframe : bool
+        - dataframe_output : pandas.DataFrame
+          Must include columns: 'turbine_flow_cfs', 'efficiency', 'power_kW'.
+
+    Returns
+    -------
+    matplotlib.figure.Figure or None
+        Figure with efficiency (left y-axis) and power (right y-axis) vs flow (cfs).
+        Returns None and prints a message if `x.pandas_dataframe` is False.
+
+    Notes
+    -----
+    - Points are deduplicated by 'turbine_flow_cfs' and sorted before plotting.
+    - The function closes the figure (`plt.close()`) and returns the handle.
+    """    
+    
     # extract dataframe output, order and subset wanted values
     if x.pandas_dataframe:
         df = x.dataframe_output
@@ -52,7 +103,29 @@ def flow_efficiency_power_plot(x) :
         print('Plotting turbine efficiency annd power requires flow time series data '\
               'provided as pandas dataframe')
 
-def flow_efficiency_plot(x) :
+def flow_efficiency_plot(x):
+    """
+    Plot turbine efficiency as a function of discrete turbine flow.
+
+    Parameters
+    ----------
+    x : object
+        HydroGenerate results container with:
+        - pandas_dataframe : bool
+        - dataframe_output : pandas.DataFrame
+          Must include columns: 'turbine_flow_cfs', 'efficiency', 'power_kW'.
+
+    Returns
+    -------
+    matplotlib.figure.Figure or None
+        Figure with efficiency vs flow (cfs).
+        Returns None and prints a message if `x.pandas_dataframe` is False.
+
+    Notes
+    -----
+    Points are deduplicated by 'turbine_flow_cfs' and sorted before plotting.
+    """
+
     # extract dataframe output, order and subset wanted values
     if x.pandas_dataframe:
         df = x.dataframe_output
[…]

[…]
