### 2025-10-29 701407d
Added Additional Tests
This commit implements additional tests
