+++ tests/ideas_to_add.py
@@ -1,13 +1,31 @@
-# from HydroGenerate.utils.api_call import *
+# Test file
+
+# tests/test_hydrogenerate_fullsuite.py
+"""
+Comprehensive HydroGenerate test suite.
+
+Covers:
+- Basic, Diversion, and Hydrokinetic hydropower calculations
+- Turbine selection and efficiency curves
+- Unit conversions
+- Headloss methods
+- Revenue calculations
+- Maintenance logic
+- Error handling
+"""
+
+import numpy as np
 import pandas as pd
 import pytest
-import numpy as np
+
 from HydroGenerate.hydropower_potential import calculate_hp_potential
-from HydroGenerate.turbine_calculation import (TurbineParameters,
-                                               turbine_type_selector,
-                                               FrancisTurbine,
-                                               TurgoTurbine,
-                                               PeltonTurbine)
+from HydroGenerate.turbine_calculation import (
+    TurbineParameters,
+    turbine_type_selector,
+    FrancisTurbine,
+    TurgoTurbine,
+    PeltonTurbine,
+)
 
 
 # ---------- FIXTURE ----------
@@ -20,48 +38,17 @@ def short_flow_df():
     df.index.name = "dateTime"
     return df
 
+
 # ---------- BASIC HYDROPOWER ----------
 def test_basic_hydropower_calculation():
     """Check Basic mode computes correct rated power."""
     hp = calculate_hp_potential(flow=8000, head=20, rated_power=None, system_efficiency=0.7)
     assert hp.rated_power == pytest.approx(9483, abs=1)
 
-# ---------- DIVERSION MODE ----------
-def test_diversion_calculate_potential():
-    """Diversion mode: FDC-based design flow, headloss calc, revenue on DF input."""
-
-    flow_data = {'dateTime': pd.Series(['2010-01-01 08:00:00+00:00', '2010-01-01 08:15:00+00:00',
-                                    '2010-01-01 08:30:00+00:00', '2010-01-01 08:45:00+00:00',
-                                    '2010-01-01 09:00:00+00:00', '2010-01-01 09:15:00+00:00',
-                                    '2010-01-01 09:30:00+00:00', '2010-01-01 09:45:00+00:00',
-                                    '2010-01-01 10:00:00+00:00', '2010-01-01 10:15:00+00:00']),
-            'discharge_cfs': pd.Series([3260, 3270, 3250, 3270, 3310, 3290, 3300,
-                                        3300, 3330, 3260])}
-
-    flow = pd.DataFrame(flow_data)
-    flow['dateTime'] = pd.to_datetime(flow['dateTime']) # preprocessing convert to datetime
-    flow = flow.set_index('dateTime') # set datetime index # flolw is in cfs
-
-    head = 20 # ft
-    power = None
-    penstock_length = 50 # ft
-    hp_type = 'Diversion' 
-
-    hp = calculate_hp_potential(flow= flow, rated_power= power, head= head,
-                                pctime_runfull = 30,
-                                penstock_headloss_calculation= True,
-                                design_flow= None,
-                                electricity_sell_price = 0.05,
-                                resource_category= 'CanalConduit',
-                                hydropower_type= hp_type, penstock_length= penstock_length,
-                                flow_column= 'discharge_cfs', annual_caclulation= True)
-    assert round(hp.rated_power, 0) == 4505
-
-
 
 # ---------- DIVERSION MODE ----------
-def test_diversion_(short_flow_df):
[…]
