+++ docs/_config.yml
@@ -1,0 +1,34 @@
+# Book settings
+# Learn more at https://jupyterbook.org/customize/config.html
+
+title: 
+author: Idaho National Laboratory
+logo: HG_horizontal.png
+
+# Force re-execution of notebooks on each build.
+# See https://jupyterbook.org/content/execute.html
+execute:
+  execute_notebooks: force
+
+# Define the name of the latex output file for PDF builds
+#latex:
+#  latex_documents:
+#    targetname: book.tex
+
+# Add a bibtex file so that we can create citations
+bibtex_bibfiles:
+  - references.bib
+
+# Information about where the book exists on the web
+repository:
+  url: https://github.com/IdahoLabResearch/HydroGenerate
+  #path_to_book: docs
+  branch: jugc-docs
+
+# Add GitHub buttons to your book
+# See https://jupyterbook.org/customize/config.html#add-a-link-to-your-repository
+html:
+  favicon: HG_icon.ico
+  baseurl: https://idaholabresearch.github.io/HydroGenerate/
+  use_issues_button: true
+  use_repository_button: true
