+++ docs/_config.yml
@@ -4,11 +4,18 @@
 title: 
 author: Idaho National Laboratory
 logo: HG_horizontal.png
+only_build_toc_files: True
 
 # Force re-execution of notebooks on each build.
 # See https://jupyterbook.org/content/execute.html
 execute:
   execute_notebooks: force
+
+sphinx:
+  extra_extensions:
+  - 'sphinx.ext.autodoc'
+  - 'sphinx.ext.napoleon'
+  - 'sphinx.ext.viewcode'
 
 # Define the name of the latex output file for PDF builds
 #latex:
