+++ tests/test_ideas.py
@@ -1,223 +1,0 @@
-# Test file
-
-# tests/test_hydrogenerate_fullsuite.py
-"""
-Comprehensive HydroGenerate test suite.
-
-Covers:
-- Basic, Diversion, and Hydrokinetic hydropower calculations
-- Turbine selection and efficiency curves
-- Unit conversions
-- Headloss methods
-- Revenue calculations
-- Maintenance logic
-- Error handling
-"""
-
-import numpy as np
-import pandas as pd
-import pytest
-
-from HydroGenerate.hydropower_potential import calculate_hp_potential
-from HydroGenerate.turbine_calculation import (
-    TurbineParameters,
-    turbine_type_selector,
-    FrancisTurbine,
-    TurgoTurbine,
-    PeltonTurbine,
-)
-
-
-# ---------- FIXTURE ----------
-@pytest.fixture
-def short_flow_df():
-    """Short time-series DataFrame for Diversion tests."""
-    rng = pd.date_range("2010-01-01", periods=4 * 24 * 40, freq="15min", tz="UTC")
-    flow = 3000 + 200 * np.sin(np.linspace(0, 20 * np.pi, len(rng)))
-    df = pd.DataFrame({"discharge_cfs": flow}, index=rng)
-    df.index.name = "dateTime"
-    return df
-
-
-# ---------- BASIC HYDROPOWER ----------
-def test_basic_hydropower_calculation():
-    """Check Basic mode computes correct rated power."""
-    hp = calculate_hp_potential(flow=8000, head=20, rated_power=None, system_efficiency=0.7)
-    assert hp.rated_power == pytest.approx(9483, abs=1)
-
-
-# ---------- DIVERSION MODE ----------
-def test_diversion_full_pipeline_with_maintenance(short_flow_df):
-    """Diversion mode full pipeline with maintenance and headloss."""
-    hp = calculate_hp_potential(
-        flow=short_flow_df,
-        flow_column="discharge_cfs",
-        head=20,
-        units="US",
-        hydropower_type="Diversion",
-        penstock_headloss_calculation=True,
-        penstock_headloss_method="Darcy-Weisbach",
-        penstock_length=50,
-        pctime_runfull=30,
-        annual_maintenance_flag=True,
-        major_maintenance_flag=True,
-        annual_caclulation=True,
-        electricity_sell_price=0.05,
-        resource_category="CanalConduit",
-    )
-    assert hp.rated_power > 0
-    assert (hp.turbine_efficiency >= 0).all() and (hp.turbine_efficiency <= 1).all()
-    assert hp.head_loss is not None
-
-
-# ---------- UNITS ROUND-TRIP ----------
-def test_units_roundtrip_basic():
-    """US vs SI unit consistency."""
-    hp_us = calculate_hp_potential(flow=8000, head=20, units="US", system_efficiency=0.7)
-    hp_si = calculate_hp_potential(flow=8000 * 0.0283168, head=20 * 0.3048, units="SI", system_efficiency=0.7)
-    assert hp_us.rated_power == pytest.approx(hp_si.rated_power, rel=1e-3)
-
-
-# ---------- HEADLOSS METHODS ----------
-@pytest.mark.parametrize("method", ["Darcy-Weisbach", "Hazen-Williams"])
-def test_headloss_methods_run(method):
-    """Both headloss methods execute successfully."""
-    hp = calculate_hp_potential(
-        flow=6000,
-        head=30,
-        units="US",
-        hydropower_type="Diversion",
-        penstock_headloss_calculation=True,
-        penstock_headloss_method=method,
-        penstock_length=100,
-    )
-    assert hp.head_loss is not None
-
-
-# ---------- TURBINE SELECTION ----------
[…]
