+++ _sources/api/hydropower_potential.rst
@@ -1,0 +1,11 @@
+HydroGenerate.hydropower_potential
+====================================
+
+Below is an auto-generated reference from the package docstring.
+
+.. automodule::
+    hydropower_potential
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
