+++ _sources/api/turbine_calculation.rst
@@ -1,0 +1,14 @@
+HydroGenerate.turbine_calculation
+========================================
+
+Below is an auto-generated reference from the package docstring.
+
+
+.. automodule::
+HydroGenerate.turbine_calculation
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+    
+
