+++ tests/test_hg.py
@@ -2,12 +2,13 @@
 import pandas as pd
 import pytest
 import numpy as np
-from HydroGenerate.hydropower_potential import calculate_hp_potential
+from HydroGenerate.hydropower_potential import calculate_hp_potential, calculate_head
 from HydroGenerate.turbine_calculation import (TurbineParameters,
                                                turbine_type_selector,
                                                FrancisTurbine,
                                                TurgoTurbine,
                                                PeltonTurbine)
+from HydroGenerate.flow_preprocessing import FlowPreProcessing
 
 
 # ---------- FIXTURE ----------
@@ -147,4 +148,274 @@ def test_invalid_units_raise():
         calculate_hp_potential(flow=1000, head=10, units="METRIC")
 
 
+
+def _df_from(vals, freq="h"): 
+    idx = pd.date_range("2012-05-01 00:00:00", periods=len(vals), freq=freq, tz="UTC")
+    df = pd.DataFrame({"discharge_cfs": vals}, index=idx)
+    df.index.name = "dateTime"
+    return df
+
+
+
+def test_percent_exceedance_design_flow_uses_P70_when_pe_30():
+    """With pe=30, PercentExceedance selects the 70th percentile of the flow series."""
+    vals = np.array([1000, 1200, 1500, 1800, 2000, 2200, 2500, 3000, 3500, 4000], dtype=float)
+    df = _df_from(vals, freq="h")
+    hp = calculate_hp_potential(
+        flow=df, head=50, units="US", hydropower_type="Diversion",
+        flow_column="discharge_cfs", pctime_runfull=30,  # 30% => design flow = P70
+        penstock_headloss_calculation=False
+    )
+    expected_p70 = np.percentile(vals, 70)
+    assert hp.design_flow == pytest.approx(expected_p70, rel=1e-6)
+
+def test_min_turbineflow_default_10_percent_clamps_to_zero():
+    """When min flow not provided, default = 10% of design; flows below clamp to zero."""
+    class Obj: pass
+    o = Obj()
+    o.flow = np.array([2.0, 8.0, 10.0, 15.0, 25.0])  # cfs
+    o.design_flow = 100.0
+    o.minimum_turbineflow = None
+    o.minimum_turbineflow_percent = None
+    FlowPreProcessing().max_turbineflow_checker(o)   # creates turbine_flow
+    FlowPreProcessing().min_turbineflow_checker(o)   # apply min clamp
+    assert np.array_equal(o.turbine_flow, np.array([0.0, 0.0, 10.0, 15.0, 25.0]))
+
+def test_darcy_headloss_increases_with_length_si():
+    """Longer penstock length => larger headloss => lower rated power (SI)."""
+    hp_short = calculate_hp_potential(
+        flow=1.5, head=60.0, units="SI", hydropower_type="Diversion",
+        penstock_headloss_calculation=True, penstock_headloss_method="Darcy-Weisbach",
+        penstock_length=50.0, penstock_material="Steel",
+        design_flow=1.5, pctime_runfull=30
+    )
+    hp_long = calculate_hp_potential(
+        flow=1.5, head=60.0, units="SI", hydropower_type="Diversion",
+        penstock_headloss_calculation=True, penstock_headloss_method="Darcy-Weisbach",
+        penstock_length=150.0, penstock_material="Steel",
+        design_flow=1.5, pctime_runfull=30
+    )
+    assert hp_long.rated_power < hp_short.rated_power
+
+def test_hazen_williams_infers_diameter_when_missing():
+    """HW method should compute diameter if not provided when max_headloss_allowed present."""
+    hp = calculate_hp_potential(
[…]
