### 2025-10-29 916c6be
Added install test option
