### 2025-11-04 #8972 @antviro
I am reviewing it the response seems right. I wanted to check the examples by my self and I have just installed the software. And I have found this error:
`
Python 3.11.2 (main, Apr 28 2025, 14:11:48) [GCC 12.2.0]
Type 'copyright', 'credits' or 'license' for more information
IPython 8.5.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: from HydroGenerate.hydropower_potential import calculate_hp_potential
Cell In [1], line 1
----> 1 from HydroGenerate.hydropower_potential import calculate_hp_potential

File ~/virtual_environments_python/hydrogenerate/lib/python3.11/site-packages/HydroGenerate/hydropower_potential.py:15
     13 import math
     14 from numpy.core.fromnumeric import mean
---> 15 from numpy.lib.function_base import median
     16 import pandas as pd
     17 from datetime import datetime

ModuleNotFoundError: No module named 'numpy.lib.function_base'

`
I know I am using python 3.11 instead of 3.10 (I have actually install 3.9 and 3.11) but I don't think it should be related. What do you think?
