### 2025-12-03 #8972 @PennyHow
Hi @snag427 and co-authors, I've just looked at the HydroGenerate repository and package again, and it is greatly improved upon - the package testing and documentation in particular. Thanks for taking the time to reply to the comments and implementing this feedback into HydroGenerate.

I just have a couple of things that need resolving still:

- I want to follow up on the auto-doc genetration from docstrings for the package, as you say that you are still working on it in your .docx response. Can we please have an update on this?

- On the documentation [Installation page](https://idaholabresearch.github.io/HydroGenerate/GettingStarted_1_Installation.html), the optional unit testing should be part of the developer install instead of after basic usage (as the unit tests can only be run from the repository and not the pip package)
- For the time-series example in [Getting Started](https://idaholabresearch.github.io/HydroGenerate/GettingStarted_2_QuickStart.html), the example file `examples/test_flow.csv` is not provided in the repo. I tried to run the example with `examples/data_test.csv` instead and came across a bug in the subsequent hp potential calculation:

```python
>>> hp = calculate_hp_potential(flow=flow_df.copy(),
...                             rated_power= power,
...                             head= head,
...                             hydropower_type= 'Diversion',
...                             flow_column= 'discharge_cfs',
...                             annual_caclulation= True)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/pho/python_workspace/HydroGenerate/HydroGenerate/hydropower_potential.py", line 1016, in calculate_hp_potential
    ConstantEletrictyPrice_pd().revenue_calculation(all_params, flow= flow)     # calculate revenue when a time-indexed pd.dataframe is given.
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/pho/python_workspace/HydroGenerate/HydroGenerate/hydropower_potential.py", line 777, in revenue_calculation
    hours = flow.index.to_series().diff().values / pd.Timedelta('1 hour')       # time difference in hours
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~
  File "pandas/_libs/tslibs/timedeltas.pyx", line 2104, in pandas._libs.tslibs.timedeltas.Timedelta.__rtruediv__
numpy._core._exceptions._UFuncBinaryResolutionError: ufunc 'divide' cannot use operands with types dtype('float64') and dtype('<m8[ns]')
```
A similar error occurs for the hp potential calculations in the [Cost and Revenue Calcutions](https://idaholabresearch.github.io/HydroGenerate/UserGuide_5_Cost%26Revenue.html). I'll include an example below from example 1 (Cost and Annual O&M as part of a project):

```python
>>> hp = calculate_hp_potential(flow= flow, rated_power= power, head= head,
...                             penstock_headloss_calculation= True,
...                             hydropower_type= hydropower_type, penstock_length = penstock_length,
...                             resource_category= resource_category,
...                             capacity_factor = capacity_factor,
...                             annual_caclulation = True)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/pho/python_workspace/HydroGenerate/HydroGenerate/hydropower_potential.py", line 1008, in calculate_hp_potential
    ONRL_BaselineCostModeling_V2().cost_calculation(all_params)
  File "/home/pho/python_workspace/HydroGenerate/HydroGenerate/hydropower_potential.py", line 662, in cost_calculation
    icc = icc / 1000000     # icc, million $
          ^^^
UnboundLocalError: cannot access local variable 'icc' where it is not associated with a value
```

I've tested this in Python 3.10 and 3.11 environments and get the same errors, so I believe this is something that needs to be resolved from your side. I'm happy to provide more information and de-bug this together if needed.
