+++ docs/Theory_1_Hydropower.md
@@ -1,79 +1,0 @@
-(hydropower-eq)= 
-## Hydropower Estimation
-Hydrogenerate canculates hydropower potential using 
-
-$$P = \eta \times \gamma \times Q \times H$$
-
-where,  
-
-* $P$ is the hydropower potential (watt),  
-* $\eta$ is the overall system efficiency (dimensionless),  
-* $\gamma$ is the specific weight of water (9,810 N/m^3), 
-* $Q$ is the flow (m^3/s), and  
-* $H$ is the net hydraulic head (m). 
-
-HG uses a simplified configuration, consisting of one penstock, turbine, and generator. The turbine type is selected based on head and design flow. Penstock, turbine diamter and other element are sized for a [design flow](Theory_2_Design_Flow.md) that can be entered by the user or computed in HG, using flow data. The overall efficiency of the system is calculated as $\eta = e_{turb}*e_{gen}$, where  $e_{turb}$  and  $e_{gen}$ are turbine and generator efficiencies. Turbine efficiencies at flows below or above the design flow are computed using a set of [empirical equations](Theory_3_Efficiency.md) for each turbine type (CANMET, 2004), while considering constant generator efficiency. 
-
-HG considers the streamflow provided by the user to be the flow routed through the hydropower plant. In some instances, e.g., multi-purpose operation reservoir, the amount of flow routed through the power system, and its operation, obey different priorities and must be analyzed outside HG to compute the flow available for hydropower generation. The considerations and analyses needed to generate the available-for-hydropower flow are outside the scope of the current version of HG. 
-
-### HG Workflow
-HG can currently compute hydropower under three hydropower types: Basic, Diversion, and Hydrokinetic. The [User Guide section](UserGuide_1_IntroductionTutorial.ipynb) includes multiple examples that illustrate each of these hydropwoer types as well as additional funcitionality. 
-
-### Basic hydropower calculation:
-Required inputs: Hydraulic head, flow, rated power. In the basic calculation mode, users can input two of these three variables and HG will calculate the third. 
-
-### Diversion
-Required inputs: Hydraulic head, flow.
-Optional inputs: all the inputs listed in Table 1 in the [User Guide section](UserGuide_7_HydroGenerateWorkflow.md), except those used only for hydrokinetic estimations.
-
-### Computation Steps: 
-1. Compute head losses using the method described in the [Theory Section - Head Loss Estimation](Theory_4_HeadLoss.md).
-2. Calculate the design flow as described in the [Theory Section - Design Flow Estimation](Theory_2_Design_Flow.md) ($Q_d$).
-3. Select the turbine type based on the net hydraulic head and design flow folowing the graph included in the [Theory Section - Turbine Selection](Theory_3_Efficiency.md).
-4. Calculate turbine efficiency using the corresponding turbine equation included in the [Theory Section - Turbine Efficiency](Theory_3_Efficiency.md) for flows above and below the design flow. Turbine efficiency is calculated following the methodology presented in the Clean Energy Project Analysis electronic textbook (CANMET Energy Technology Center, 2004) 
-5. Estimate estimate power for the value (s) of flow provided the user implementing any contraint selected. Flow constraints are discussed in the [Theory Section - Flow Constraints](Theory_6_FlowConstraints.md)
-6. Compute the overall efficiency of the system. The overall efficiency of the system is computed as the turbine efficiency (calculated in step 4) times the Generator efficiency. 
[…]

[…]
