### 2025-12-11 8c00d82
Updated _toc.yml
made minor syntax corrections
