### 2025-12-11 bcf0c95
Added missing documentation files
