### 2025-12-18 1f396d9
Building Book after SN update
