+++ docs/api/FPP.md
@@ -1,0 +1,10 @@
+
+## Flow Preprocessing API Reference
+
+This page documents the classes and functions in the `flow_preprocessing` module of the HydroGenerate package.
+
+```{eval-rst}
+.. automodule:: HydroGenerate.flow_preprocessing
+   :members:
+   :undoc-members:
+   :show-inheritance:
