### 2026-01-08 6182b08
Modifications required for docstrings documentation
1. Files under HydroGenerate folder have modified docstrings so that they are sphinx compatible. Main code has not been modified.

2. The doc folder is updated with the new api folder, which has sphinx compatible markdowns for each module.
