+++ HydroGenerate/hydraulic_processing.py
@@ -10,7 +10,6 @@ Functionality
 - Size penstocks to respect a maximum allowable head loss.
 """
 
-
 import numpy as np
 from abc import ABC, abstractmethod
 
@@ -19,61 +18,65 @@ nu = 0.0000011223       # Kinematic viscosity of water (m2/s) at 60F (~15.6C) - 
 
 # Hydraulic design parameters
 class HydraulicDesignParameters:
-    """
-    Container for hydraulic inputs/derived values used in head-loss and sizing.
+
+    """Hydraulic design parameter container.
+
+    This class stores input values needed for penstock head-loss calculations
+    and provides helper methods for computing Reynolds number and a starting
+    design diameter for iterative sizing.
 
     Parameters
     ----------
-    flow : float or numpy.ndarray
-        Flow values (m^3/s) for general context or time-series use.
+    flow : float or array-like
+        Flow values (m³/s). This may represent a single value or a time series,
+        depending on the calling context.
     design_flow : float
-        Design discharge (m^3/s) used for diameter/head-loss sizing.
-    head : float or None
-        Available gross head (m).
+        Design flow (m³/s).
+    head : float
+        Available (gross) head (m).
     head_loss : float or numpy.ndarray or None
-        Head loss (m). Can be precomputed or filled by calculators.
+        Head loss for the provided ``flow`` values (m). This may be provided by
+        the user or computed by head-loss calculators in this module.
     max_headloss_allowed : float or None
-        Maximum head loss as a percent of `head` (1–100). If None, 10% default
-        is applied in Darcy–Weisbach sizing.
+        Maximum head loss allowed as a percentage of head (%, 1–100). If
+        ``None``, method-specific defaults may be applied.
     penstock_headloss_method : str or None
-        Text identifier for method: e.g., "Darcy-Weisbach" or "Hazen-Williams".
+        Name of the head-loss method. Current options in this module include
+        ``"Darcy-Weisbach"`` and ``"Hazen-Williams"``. (Manning is referenced
+        but not implemented in this file.)
     penstock_headloss_calculation : bool
-        If True, compute penstock head loss; set False when net head is given.
+        Whether to calculate head loss in the penstock. If net head is already
+        available externally, set to ``False``.
     penstock_length : float or None
-        Penstock length (m).
+        Penstock length (m). Required for head-loss calculations that depend on
+        length.
     penstock_diameter : float or None
-        Known penstock diameter (m). If None, methods will size it.
+        Penstock diameter (m). If ``None``, some methods may size diameter to
+        meet the allowed head-loss constraint.
     penstock_material : str or None
-        Material key in roughness tables (e.g., "Steel", "Concrete", etc.).
+        Penstock material key used for roughness selection. Expected options include: ``CastIron``, ``Concrete``, ``GalvanizedIron``, ``Plastic``, ``Steel``.
     penstock_frictionfactor : float or None
-        For Darcy–Weisbach: friction factor `f`. For Hazen–Williams: C-factor.
-        If None, it is computed/selected.
+        Friction parameter for the chosen method:
+        - Darcy–Weisbach: friction factor ``f`` (if provided), or roughness
+          ``epsilon`` used to compute ``f``.
+        - Hazen–Williams: coefficient ``C``.
+        - Manning: ``n`` (not used in this file).
     channel_average_velocity : float or None
-        Average channel velocity (m/s), if used elsewhere.
+        Average cross-sectional water velocity in the channel (Q/A), if
+        applicable.
 
-    Attributes
-    ----------
-    Re : list or float
-        Reynolds number (dimensionless), computed for design flow.
-    penstock_design_diameter : float or None
-        Working/selected design diameter (m).
[…]
