+++ HydroGenerate/flow_preprocessing.py
@@ -67,15 +67,25 @@ class FlowPreProcessingParameters:
     Attributes
     ----------
     minimum_turbineflow : float or None
-    minimum_turbineflow_percent : float or None
+        Absolute minimum turbine flow (m³/s). If provided, this overrides the
+        percentage-based minimum flow. If ``None``, the minimum is computed from
+        ``minimum_turbineflow_percent`` (or a default of 10%).
+    minimum_turbineflow_percent : float or int or None
+        Minimum turbine flow as a percentage of ``design_flow`` (1–100). Used
+        only when ``minimum_turbineflow`` is ``None``.
     annual_maintenance_flag : bool
+        If ``True``, apply annual maintenance by setting turbine flow to zero
+        for the lowest-flow calendar week each year (based on mean weekly flow).
     major_maintenance_flag : bool
+        If ``True``, apply major maintenance by setting turbine flow to zero
+        for the lowest-flow two-week period every five years.
 
     Notes
     -----
-    This class only stores configuration; calculations are performed by
-    methods in `FlowPreProcessing`.
-    """        
+    These parameters are typically attached to a higher-level *flow object*
+    (e.g., a project-specific container) and then read by methods in
+    :class:`~FlowPreProcessing`.
+    """      
 
     def __init__(self, minimum_turbineflow, minimum_turbineflow_percent,       # Flow parameters
                  annual_maintenance_flag, major_maintenance_flag): 
@@ -89,89 +99,49 @@ class FlowPreProcessingParameters:
 
 
 class FlowPreProcessing():
+    """Flow preprocessing routines operating on a provided flow object.
 
+    Methods in this class update values on ``flow_obj`` in-place (most
+    importantly ``flow_obj.turbine_flow``).
+
+    Notes
+    -----
+    The caller is responsible for ensuring that:
+
+    - ``flow_obj.turbine_flow`` is a 1D numpy array aligned with
+      ``flow_obj.datetime_index``.
+    - ``flow_obj.datetime_index`` is a pandas.DatetimeIndex whose name (or level
+      name) is ``"dateTime"`` so that ``pd.Grouper(level="dateTime", ...)``
+      works as written in the implementation.
     """
-    Apply design flow limits and scheduled maintenance to a flow object.
-
-    The methods in this class expect a `flow_obj` with, at minimum, the
-    attributes described in the module docstring. Methods update
-    `flow_obj.turbine_flow` in place and do not return values.
-
-    See Also
-    --------
-    max_turbineflow_checker : Cap turbine flow at design flow.
-    min_turbineflow_checker : Enforce minimum turbine flow threshold.
-    annual_maintenance_implementer : Zero 7 days/year at lowest-flow week.
-    major_maintenance_implementer : Zero 14 days/5 years at lowest bi-weekly mean.
-    """
-
-    # Function that sets max turbine flow to the design flow
-    def max_turbineflow_checker(self, flow_obj):
-        """
-        Cap turbine flow at the design flow.
-
-        Parameters
-        ----------
-        flow_obj : object
-            Object with attributes:
-            - flow : numpy.ndarray
-                Raw inflow time series (m^3/s).
-            - design_flow : float or numpy.ndarray
-                Turbine design flow (m^3/s). If array, must align with `flow`.
-
-        Side Effects
-        ------------
-        Updates `flow_obj.turbine_flow` (numpy.ndarray), where each element is
-        `min(flow, design_flow)`.
-
-        Examples
-        --------
-        >>> # flow_obj.flow = [12, 8], design_flow = 10
-        >>> # turbine_flow becomes [10, 8]
-        """
-
-        flow = flow_obj.flow            # this is a numpy array
-        design_flow = flow_obj.design_flow
[…]
