+++ docs/_config.yml
@@ -21,6 +21,12 @@ sphinx:
    - 'sphinxcontrib.autodoc_pydantic'
    - 'sphinxcontrib.mermaid'
    - 'sphinx.ext.mathjax'
+  
+  config:
+    autodoc_member_order: bysource
+    autoclass_content: both
+    add_module_names: false
+
 
 # Define the name of the latex output file for PDF builds
 #latex:
@@ -44,14 +50,3 @@ html:
   baseurl: https://idaholabresearch.github.io/HydroGenerate/
   use_issues_button: true
   use_repository_button: true
-
-sphinx:
-  extra_extensions:
-   - 'sphinx.ext.autodoc'
-   - 'sphinx.ext.autosummary'
-   - 'sphinx.ext.coverage'
-   - 'sphinx.ext.napoleon'
-   - 'sphinx.ext.intersphinx'
-   - 'sphinxcontrib.autodoc_pydantic'
-   - 'sphinxcontrib.mermaid'
-   - 'sphinx.ext.mathjax'

[…]
