+++ _sources/api/TC.md
@@ -1,0 +1,11 @@
+
+# Turbine Calculation
+
+This page documents the classes and functions in the `turbine_calculation` module of the HydroGenerate package.
+
+```{eval-rst}
+.. automodule:: HydroGenerate.turbine_calculation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: DesignFlow, Turbine
