### 2026-01-20 3ac913d
Update for docstrings documentations
1. reviewer comment on icc was because of syntax of the key "newstream-reach" which has been corrected.
2. All books have been compiled without errors
3. All images have been compiled
4. error due to max_flow_checker has been resolved
