### 2026-01-21 #8972 @snag427
Hi All,

Thank you so much for keeping this open for us. Your reviews have significantly helped us to improve the package and its documentation. We sincerely apologize for the delay in responding. Here are some updates we have made:

**Comment: I want to follow up on the auto-doc generation from docstrings for the package, as you say that you are still working on it in your .docx response. Can we please have an update on this?**

Response:  auto-doc generation from docstrings has been completed

**Comment: On the documentation [Installation page](https://idaholabresearch.github.io/HydroGenerate/GettingStarted_1_Installation.html), the optional unit testing should be part of the developer install instead of after basic usage (as the unit tests can only be run from the repository and not the pip package)**

Response: optional unit testing has been moved to the developers section

**Comment: For the time-series example in [Getting Started](https://idaholabresearch.github.io/HydroGenerate/GettingStarted_2_QuickStart.html), the example file examples/test_flow.csv is not provided in the repo. I tried to run the example with examples/data_test.csv instead and came across a bug in the subsequent hp potential calculation**

**A similar error occurs for the hp potential calculations in the [Cost and Revenue Calcutions](https://idaholabresearch.github.io/HydroGenerate/UserGuide_5_Cost%26Revenue.html). I'll include an example below from example 1 (Cost and Annual O&M as part of a project):**

Response: The above errors have been corrected.

Thank you!
