### 2026-02-16 a87fac3
Fix typos in introduction
Corrected typos and improved clarity in the introduction section.
