### 2026-02-16 10dd267
Remove http from dois in paper.bib
Updated URLs to DOIs for several references and removed some URLs.
