+++ paper.bib
@@ -4,7 +4,6 @@
   institution  = {Oak Ridge National Lab. (ORNL), Oak Ridge, TN (United States)},
   annote       = {Recent resource assessments conducted by the United States Department of Energy have identified significant opportunities for expanding hydropower generation through the addition of power to non-powered dams and on undeveloped stream-reaches. Additional interest exists in the powering of existing water resource infrastructure such as conduits and canals, upgrading and expanding existing hydropower facilities, and the construction new pumped storage hydropower. Understanding the potential future role of these hydropower resources in the nation’s energy system requires an assessment of the environmental and techno-economic issues associated with expanding hydropower generation. To facilitate these assessments, this report seeks to fill the current gaps in publically available hydropower cost estimating tools that can support the national-scale evaluation of hydropower resources.},
   doi          = {10.2172/1244193},
-  url          = {https://www.osti.gov/biblio/1244193},
   place        = {United States},
   year         = {2015},
   month        = {09}}
@@ -37,7 +36,7 @@
 
 @misc{USGS,
     title = {dataretrieval (Python): a Python package for discovering and retrieving water data available from U.S. federal hydrologic web services: U.S. Geological Survey software release},
-    url = {https://doi.org/10.5066/P94I5TX3},
+    doi = {10.5066/P94I5TX3},
     author = {U.S. Geological Survey},
     year = {2024},
     note = {Accessed on 02-18-2025}
@@ -51,8 +50,7 @@ number = {6},
 pages = {2729-2735},
 year = {2011},
 issn = {1364-0321},
-doi = {https://doi.org/10.1016/j.rser.2011.03.029},
-url = {https://www.sciencedirect.com/science/article/pii/S1364032111001328},
+doi = {10.1016/j.rser.2011.03.029},
 author = {C. Alonso-Tristán and D. González-Peña and M. Díez-Mediavilla and M. Rodríguez-Amigo and T. García-Calderón},
 keywords = {Feasibility study, Small hydropower, RETScreen, Economic study},
 abstract = {A small hydropower plant in Spain is studied from an energetic and economic perspective. The viability of the facility is examined using the freeware software RETScreen. Calculated and standard operational data are compared, thereby demonstrating the feasibility of the project from all points of view. The study highlights the growing interest in renewable energies.}
@@ -80,7 +78,6 @@ abstract = {A small hydropower plant in Spain is studied from an energetic and e
   publisher    = {Zenodo},
   version      = {2.0.7},
   doi          = {10.5281/zenodo.14776272},
-  url          = {https://doi.org/10.5281/zenodo.14776272},
   swhid        = {swh:1:dir:9bc7da99727538708857e1e27ea6542920db0fd1
                    ;origin=https://doi.org/10.5281/zenodo.5597138;vis
                    it=swh:1:snp:44de75f53264cc3b455a9d033e92649a4e751
@@ -96,8 +93,7 @@ volume = {113},
 pages = {109240},
 year = {2019},
 issn = {1364-0321},
-doi = {https://doi.org/10.1016/j.rser.2019.06.047},
-url = {https://www.sciencedirect.com/science/article/pii/S136403211930440X},
+doi = {10.1016/j.rser.2019.06.047},
 author = {C.M. Niebuhr and M. {van Dijk} and V.S. Neary and J.N. Bhagwan},
 keywords = {Hydrokinetic energy conversion, Low head hydropower, Hydrokinetic technology, Hydrokinetic applicability, Hydrokinetic turbine},
[…]
