### 2026-02-16 e0a3931
Editorial recommended changes to JOSS paper
