### 2026-02-16 #8972 @AdamRJensen
@snag427 I see that [bmitrauncc](https://github.com/bmitrauncc) has a significant number of commits (although many trivial). Please confirm that you have considered authorship and concluded that they should not be an author (leaving a thumbs up is sufficient).
