### 2026-02-16 #8972 @antviro
> [@PennyHow](https://github.com/PennyHow), [@antviro](https://github.com/antviro) It seems that [@snag427](https://github.com/snag427) has addressed many of your comments. Could you update your review checklist accordingly?

In my opinion, all comments are issues have been addressed and I consider it is ready for publication.
