### 2026-02-16 #8972 @snag427
> [@snag427](https://github.com/snag427) I see a number of open issues from PennyHow in the repository. Once you've addressed these, I'll ping the reviewers again.

I will get these cleared.
