### 2026-02-18 #8972 @snag427
> [@snag427](https://github.com/snag427) I see that [bmitrauncc](https://github.com/bmitrauncc) has a significant number of commits (although many trivial). Please confirm that you have considered authorship and concluded that they should not be an author (leaving a thumbs up is sufficient).

He made significant initial contributions so we are adding him to the paper. His workplace is different and has procedures for internal review. He is currently reviewing and will let us know shortly.
