### 2026-02-18 e60c0d9
Added Bhaskar Mitra to the paper
