### 2026-02-22 afdcb89
Changes for the paper
1. python version in setup.py has been changed from 3.8 to 3.10

2. removed https from dois

3. changed citation for readme including Soumyadeep Nag and removing version 3.6 from title
