+++ paper.md
@@ -18,6 +18,7 @@ authors:
     corresponding: true 
     affiliation: 1
   - name: Bhaskar Mitra
+    orcid: 0000-0001-5584-1390
     equal-contrib: true 
     affiliation: 2
 affiliations:
