### 2026-02-27 077019e
Fix three critical bugs: NameError, CSV overwrite, and SSL verification
- Fix NameError: PropellorTurbine -> PropellerTurbine in turbine_calculation.py
- Fix CSV overwrite: move to_csv outside loop and concat all DataFrames before
  saving; also respect the save_csv flag that was previously ignored
- Fix SSL: remove verify=False from requests.get in api_call.py

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
