### 2026-02-27 a229d22
Add type hints to all public APIs and parameter classes
- Added Optional, Union, Any, Tuple imports to all modules
- Annotated all __init__ methods on parameter classes
  (TurbineParameters, HydraulicDesignParameters,
   FlowPreProcessingParameters, EconomicParameters)
- Annotated all turbine calculator, head loss, and roughness methods
- Annotated public utility functions (calculate_hp_potential,
  pd_checker, calculate_head, get_data, format_api_response, clean_data)
- Annotated all summary/plot functions with Optional[Figure] return types
- Used Any for merged parameter objects where dynami
[…]
