+++ HydroGenerate/hydraulic_processing.py
@@ -8,6 +8,8 @@ Copyright 2023, Battelle Energy Alliance, LLC
 This script handles head loss calculations and penstock sizing 
 """
 
+from typing import Optional, Union
+
 import numpy as np
 from abc import ABC, abstractmethod
 
@@ -17,11 +19,19 @@ nu = 0.0000011223       # Kinematic viscosity of water (m2/s) at 60F (~15.6C) - 
 # Hydraulic design parameters
 class HydraulicDesignParameters:
 
-    def __init__(self, flow, design_flow,       # Flow parameters
-                 head, head_loss, 
-                 max_headloss_allowed, penstock_headloss_method, penstock_headloss_calculation,    # head / head loss 
-                 penstock_length, penstock_diameter, penstock_material, penstock_frictionfactor,        # penstock parameters  
-                 channel_average_velocity):
+    def __init__(self,
+                 flow: Optional[Union[float, np.ndarray]],
+                 design_flow: Optional[float],
+                 head: Optional[float],
+                 head_loss: Optional[float],
+                 max_headloss_allowed: Optional[float],
+                 penstock_headloss_method: Optional[str],
+                 penstock_headloss_calculation: bool,
+                 penstock_length: Optional[float],
+                 penstock_diameter: Optional[float],
+                 penstock_material: Optional[str],
+                 penstock_frictionfactor: Optional[float],
+                 channel_average_velocity: Optional[float]) -> None:
 
         '''
         This class initializes hydraulic parameters needed for multiple calculations 
@@ -49,13 +59,13 @@ class HydraulicDesignParameters:
         self.penstock_headloss_calculation = penstock_headloss_calculation      # Calculate head_loss in the penstock? (boolean) - if net head is available set to False
 
     # Reynolds number calculation
-    def reynoldsnumber_calculator(self):
+    def reynoldsnumber_calculator(self) -> None:
 
         self.designdiameter_calculator()        # calculate diameter
         self.Re = 4 * self.design_flow / (np.pi * self.penstock_design_diameter * nu)        # Reynolds number, update  
 
     # Design diameter calculator - initial assumed diameter
-    def designdiameter_calculator(self):
+    def designdiameter_calculator(self) -> None:
         
         if self.penstock_diameter is not None:      # penstock diameter is known
             self.penstock_design_diameter = self.penstock_diameter       # update
@@ -66,7 +76,7 @@ class HydraulicDesignParameters:
 
 # Roughness coefficients
 class RoughnessCoefficients:
-    def __init__(self, material, darcyweisbach_epsilon, hazenwiliams_c, mannings_n):        # Epsilon, C, n
+    def __init__(self, material: str, darcyweisbach_epsilon: float, hazenwiliams_c: float, mannings_n: float) -> None:
         self.material = material
         self.darcyweisbach_epsilon = darcyweisbach_epsilon
         self.hazenwiliams_c = hazenwiliams_c
@@ -81,7 +91,7 @@ class RoughnessCoefficients:
             self.add_roughnesscoefficient_values("Plastic", 0.000001524, 145, 0.013)        # Plastic. Epsilon, C, n
             self.add_roughnesscoefficient_values("Steel", 0.00004572, 145, 0.016)           # Steel. Epsilon, C, n
 
-        def add_roughnesscoefficient_values(self, material, darcyweisbach_epsilon, hazenwiliams_c, mannings_n):     # method for adding roughness values
+        def add_roughnesscoefficient_values(self, material: str, darcyweisbach_epsilon: float, hazenwiliams_c: float, mannings_n: float) -> None:
[…]
