### 2026-02-27 85be186
Replace print() with logging module in summary_results and api_call
- Added module-level loggers via logging.getLogger(__name__)
- summary_results.py: 6 user-facing print warnings -> logger.warning()
  (also fixed 'annd' typo in one message)
- api_call.py: API URL -> logger.debug(), success msg -> logger.info(),
  HTTP error status -> logger.error(), site info -> logger.info()
- Print statements in __main__ blocks left unchanged (appropriate there)

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
