### 2026-02-27 b993d5c
Replace wildcard imports with explicit imports in hydropower_potential
- Replaced 'from module import *' with named imports for all three
  internal modules, listing only the symbols actually used
- Removed unused 'from HydroGenerate.summary_results import *' entirely
  (AST analysis confirmed zero symbols from it were referenced)
- Prevents namespace pollution and makes dependencies explicit

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
