### 2026-02-27 61233df
Fix numpy deprecation warning and update notebook imports
- turbine_calculation.py: fix array-to-scalar conversion by adding
  .item() before float() cast, resolving NumPy 1.25 DeprecationWarning
- DiversionMode.ipynb: replace wildcard import with explicit imports
  of calculate_hp_potential and all summary_results plot functions
  (required after removing wildcard re-export from hydropower_potential)
- BasicMode.ipynb, EconomicCalculation.ipynb, Hydrokinetics.ipynb:
  switch from 'import *' to explicit calculate_hp_potential import;
  EconomicCalculation adds explicit numpy import since it us
[…]
