### 2026-03-02 953675a
Some minor additional changes for the paper
1. test_main.yml - changed python version from 3.8 to 3.10
2. citation.cff - had to add my name
3. paper.bib - had to add my name in HG citation
4. README.md - Had to add my name
5. setup.py - changed python version from 3.8 to 3.10
