+++ paper.bib
@@ -155,7 +155,7 @@ author = {Nikolay Arefiev and Olga Nikonova and Nikolay Badenko and Timofey Ivan
 
 @misc{HG,
 title = {Hydrogenerate: Open Source Python Tool To Estimate Hydropower Generation Time-series},
-author = {Mitra, Bhaskar and Gallego-Calderon, Juan F. and Elliott, Shiloh N and Mosier, Thomas M and Bastidas Pacheco, Camilo Jose, Nag, Soumyadeep and USDOE Office of Energy Efficiency and Renewable Energy},
+author = {Mitra, Bhaskar and Gallego-Calderon, Juan F. and Elliott, Shiloh N and Mosier, Thomas M and Bastidas Pacheco, Camilo Jose and Nag, Soumyadeep and USDOE Office of Energy Efficiency and Renewable Energy},
 abstractNote = {Hydropower is one of the most mature forms of renewable energy generation. The United States (US) has almost 103 GW of installed, with 80 GW of conventional generation and 23 GW of pumped hydropower [1]. Moreover, the potential for future development on Non-Powered Dams is up to 10 GW. With the US setting its goals to become carbon neutral [2], more renewable energy in the form of hydropower needs to be integrated with the grid. Currently, there are no publicly available tool that can estimate the hydropower potential for existing hydropower dams or other non-powered dams. The HydroGenerate is an open-source python library that has the capability of estimating hydropower generation based on flow rate either provided by the user or received from United States Geological Survey (USGS) water data services. The tool calculates the efficiency as a function of flow based on the turbine type either selected by the user or estimated based on the “head” provided by the user.},
 url = {https://www.osti.gov//servlets/purl/1829986},
 doi = {10.11578/dc.20211112.1},
