+++ paper.bib
@@ -13,7 +13,7 @@
 
 @misc{EIA_price,
     title = {Wholesale Electricity and Natural Gas Market Data},
-    author = {US Energy Information Administration},
+    author = {{US Energy Information Administration}},
     year = {2025},
     url = {https://www.eia.gov/electricity/wholesale/},
     urldate      = {02/04/2025}
@@ -30,7 +30,7 @@
 @misc{irrigationviz,
     title = {irrigationviz},
     url = {https://irrigationviz.pnnl.gov/},
-    author = {Pacific Northwest Natiola Laboratory},
+    author = {{Pacific Northwest National Laboratory}},
     year = {2024},
     note = {Accessed on 02-18-2025}
 }
@@ -38,7 +38,7 @@
 @misc{USGS,
     title = {dataretrieval (Python): a Python package for discovering and retrieving water data available from U.S. federal hydrologic web services: U.S. Geological Survey software release},
     url = {10.5066/P94I5TX3},
-    author = {U.S. Geological Survey},
+    author = {USGS},
     year = {2024},
     note = {Accessed on 02-18-2025}
 }
@@ -62,7 +62,7 @@ abstract = {A small hydropower plant in Spain is studied from an energetic and e
   TITLE = {Hydropower Engineering Handbook},
   SUBTITLE = {},
   AUTHOR = {Gulliver, John S. and Arndt, Roger E.A.},
-  YEAR = {1991}, % I looked it up
+  YEAR = {1991},
   PUBLISHER = {McGraw-Hill, Inc.},
 }
 
@@ -155,7 +155,7 @@ author = {Nikolay Arefiev and Olga Nikonova and Nikolay Badenko and Timofey Ivan
 
 @misc{HG,
 title = {Hydrogenerate: Open Source Python Tool To Estimate Hydropower Generation Time-series},
-author = {Mitra, Bhaskar and Gallego-Calderon, Juan F. and Elliott, Shiloh N and Mosier, Thomas M and Bastidas Pacheco, Camilo Jose and Nag, Soumyadeep and USDOE Office of Energy Efficiency and Renewable Energy},
+author = {Mitra, Bhaskar and Gallego-Calderon, Juan F. and Elliott, Shiloh N and Mosier, Thomas M and Bastidas Pacheco, Camilo Jose and Nag, Soumyadeep and {USDOE Office of Energy Efficiency and Renewable Energy}},
 abstractNote = {Hydropower is one of the most mature forms of renewable energy generation. The United States (US) has almost 103 GW of installed, with 80 GW of conventional generation and 23 GW of pumped hydropower [1]. Moreover, the potential for future development on Non-Powered Dams is up to 10 GW. With the US setting its goals to become carbon neutral [2], more renewable energy in the form of hydropower needs to be integrated with the grid. Currently, there are no publicly available tool that can estimate the hydropower potential for existing hydropower dams or other non-powered dams. The HydroGenerate is an open-source python library that has the capability of estimating hydropower generation based on flow rate either provided by the user or received from United States Geological Survey (USGS) water data services. The tool calculates the efficiency as a function of flow based on the turbine type either selected by the user or estimated based on the “head” provided by the user.},
 url = {https://www.osti.gov//servlets/purl/1829986},
 doi = {10.11578/dc.20211112.1},
