### 2026-03-02 6baf821
Correction of version in test_main.yml
The python version 3.10 was changed to '3.10'
