### 2026-03-03 ac02ce2
Replace linear diameter search with binary search in diameter_check
The old implementation crawled from 0.01m upward in 0.1m steps, calling
penstock_headloss_calculator recursively each step — O(D/0.1) iterations
and deep call stacks for large pipes.

New approach:
1. Find upper bound by doubling diameter until head loss is within limit
   (~log2(D_final/D_start) steps)
2. Binary search between bounds to 1mm precision
   (~log2(range/0.001) steps)

For a 10m pipe this reduces ~100 iterations to ~22. The result also
changes slightly: the binary search finds the precise minimum valid
diameter ra
[…]
