### 2026-03-04 dba60b3
More minor changes to the paper based on Bhaskar
1. paper.md - Mostly grammatical errors were corrected.
2. README.md - reference error  was corrected.
