+++ paper.md
@@ -44,9 +44,9 @@ Additionally, researchers can leverage HydroGenerate for large and small-scale h
 
 Additional research has uncovered automated approaches for identifying sites with hydropower potential by using Geographic Information Systems (GIS) in combination with summaries of flow data [@Arefiev]. HydroGenerate can be used in combination with tools implementing this or other GIS functionality, to analyze generation and plant characteristics once sites have been identified and expand their usability. For example, HydroGenerate was used to assess hydropower potential across multiple non-powered dams where head data was available or was estimated using remote sensing data [@osti_1968288]. Additionally, HydroGenerate is used on IrrigationViz, a GIS-based software, to assess hydropower potential on agricultural infrastructure, demonstrating its flexibility for supporting other analysis [@irrigationviz].   
 
-# The HydroGenerate Workflow
+# The HydroGenerate workflow
 
-After the user has chosen a site for hydropower potential analysis, and collected the data needed (head, flow that can be procured from [@USGS_data] and, site details that are described in [@HG]), they can use HydroGenerate in multiple ways. In the most common scenario, users will site data, input flow and head data to evaluate hydropower potential. However, they can also use the tool to evaluate site requirements for a desired amount of power that can lead them to understand site-development requirements. Flow data can be entered as a single number or a time series such as those available from the USGS. HydroGenerate can work with flow data collected at different temporal frequencies. In addition to flow and head data, there are 30 inputs available to the user e.g., turbine type, maintenance constraints, unit system, penstock length, allowing further customization in a project evaluation [@HG]. Upon execution, HydroGenerate will create an object which contains design parameters, performance estimates, and additional data as its attributes. The workflow of HydroGenerate as displayed in Figure \ref{HG_WF}, can be divided into the following steps:
+After the user has chosen a site for hydropower potential analysis, and collected the data needed (head, flow that can be procured from [@USGS_data] and, site details that are described in [@HG]), they can use HydroGenerate in multiple ways. In the most common scenario, users will input site data, such as flow and head data to evaluate hydropower potential. However, they can also use the tool to evaluate site requirements for a desired amount of power that can lead them to understand site-development requirements. Flow data can be entered as a single number or a time series such as those available from the USGS. HydroGenerate can work with flow data collected at different temporal frequencies. In addition to flow and head data, there are 30 inputs available to the user e.g., turbine type, maintenance constraints, unit system, penstock length, allowing further customization in a project evaluation [@HG]. These inputs and following procedures have been detailed in the user guide for [@HG]. Upon execution, HydroGenerate will create an object which contains design parameters, performance estimates, and additional data as its attributes. The workflow of HydroGenerate as displayed in Figure \ref{HG_WF}, can be divided into the following steps:
 
 * Step 1 Input data processing: This step involves conversion of inputs to SI units when using US Customary Units. HydroGenerate performs all computations in SI units except for the cost models as these were developed in US Customary Units. 
 
@@ -62,7 +62,7 @@ After the user has chosen a site for hydropower potential analysis, and collecte
 
[…]
