### 2026-03-13 #8972 @AdamRJensen
@snag427 Please address https://github.com/IdahoLabResearch/HydroGenerate/pull/43
