### 2026-03-22 #8972 @snag427
Hi, The references have been corrected in the updated version of the main branch of the repository. https://github.com/IdahoLabResearch/HydroGenerate

my new entry is

@book{HPEH,
  author    = {Gulliver, John S. and Arndt, Roger E. A.},
  title     = {Hydropower Engineering Handbook},
  publisher = {McGraw-Hill},
  year      = {1991}
}

which does not have the subtitle but has the publisher fields.
