### 2026-03-23 #8972 @AdamRJensen
> Hi, The references have been corrected in the updated version of the main branch of the repository. https://github.com/IdahoLabResearch/HydroGenerate

@snag427 The paper is built from the branch "Soumyadeep_Paper" as specified in the top of this page. Please make sure that that branch has the correct files.
