### 2026-03-23 #8972 @AdamRJensen
@snag427 Could you make a new release of the package and comment here with the version number. Afterwards, please upload the source code to Zenodo and provide me with the doi.
