### 2026-07-01 #8972 @kyleniemeyer
@snag427 the issue is not with the `.bib` entries, but how you are citing in the text of the paper in a few places.

For example, in the "The HydroGenerate workflow" section, you have `site details that are described in [@HG])` which formats as
> site details that are described in (Mitra et al., 2021)

but this should be `site details that are described by @HG` (note the square brackets are gone) to get
> site details that are described by Mitra et al. (2021)

This issue appears in a few places, where the citation is used as an object in the sentence rather than like a footnote.

The other issue is related to when you have multiple citations that should be together. For example, in the "Step 2a" bullet text, you have `flow duration relationship [@ALONSOTRISTAN20112729], [@RETSCREEN], [@HPEH].`, which currently formats as
> flow duration relationship (Alonso-Tristán et al., 2011), (Canada, 2025), (Gulliver & Arndt, 1991).

This should instead be `flow duration relationship [@ALONSOTRISTAN20112729; @RETSCREEN; @HPEH].` (note that the three citations are included in a single command, separated by semicolons) which will format as
> flow duration relationship (Alonso-Tristán et al., 2011; Canada, 2025; Gulliver & Arndt, 1991).

And similarly in a few other places.
