### 2026-08-18 #8972 @snag427
Hi @kyleniemeyer , Sorry for the break in communication.

On your comment "We require the peer-reviewed version of the software to be archived in a way that ensures (that version of) the software remains accessible.", I wanted to clarify that the intention was to update and make the tool accessible through Github which is the case currently. We have been working on the previous versions to update the same. Do we still need another shared version at Zenodo or figshare? Please let me know.

Also, the DOE CODE site https://www.osti.gov/doecode/biblio/66918 has the link to the current version on Github.
