### 2026-09-06 #8972 @snag427
@kyleniemeyer Here is the Zenodo DOI
[10.5281/zenodo.22546815](https://doi.org/10.5281/zenodo.22546815)

Also, I have corrected the citation issue with
"I do see one spot that still needs the change: "head, flow that can be procured from (USGS, 2025)" near the beginning of "The HydroGenerate workflow, which should instead read "head, flow that can be procured from USGS (2025)"."

Please let me know the next steps.
